requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
requests==2.32.3
python-dotenv==1.0.1
bcrypt==4.1.3
cryptography==42.0.8
python-telegram-bot==21.4
openai==1.40.2
